requirements.txt
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==23.2.0
audioop-lts==0.2.1
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.38.24
botocore==1.38.24
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cog==0.14.10
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
daphne==4.1.2
distro==1.9.0
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-storages==1.14.6
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
gradio==5.29.0
gradio_client==1.10.0
groovy==0.1.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openai==1.75.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
replicate==1.0.4
requests==2.32.3
rich==14.0.0
ruff==0.11.9
s3transfer==0.13.0
safehttpx==0.1.6
semantic-version==2.10.0
service-identity==24.2.0
setuptools==80.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SoundCard==0.4.4
soundfile==0.13.1
sqlparse==0.5.3
starlette==0.46.2
stripe==12.1.0
structlog==24.4.0
sympy==1.14.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
trimesh==4.6.8
triton==3.3.0
Twisted==24.11.0
txaio==23.1.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
uvicorn==0.34.1
uvloop==0.21.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whisper==1.1.10
zope.interface==7.2
